Added svn pull, and save master locally functions
