+++ Misc Tools/Limits Version Control Tool/components/jira_items.py
@@ -25,13 +25,11 @@ def init_jira_connection():
 
 
 def handle_jira_error(self, error_type, error_message):
-    status, title, reporter= "ERROR", "N/A", "N/A"
-    icon, color_code= "🔴", "red"
     self.is_jira_valid= False
     self.jira_status.setText(
-        f"Jira Status: {icon}<br>Ticket: <b style='color:{color_code};'>{error_type}</b>, "
-        f"Status: <b style='color:{color_code};'>{status}</b>"
-        f"<br>Title: {title}<br>Reporter: {reporter}")
+        f"Jira Status: 🔴<br>Ticket: <b style='color:red;'>{error_type}</b>, "
+        f"Status: <b style='color:red;'>ERROR</b>"
+        f"<br>Title: N/A<br>Reporter: N/A")
     QMessageBox.information(self, "JIRA Check Failed", error_message)
 
 
@@ -40,20 +38,13 @@ def check_jira_status(self):
         jira_connection= init_jira_connection()
         self.ticket_obj= jira_connection.issue(self.jira_ticket.text())
         status= str(self.ticket_obj.fields.status).strip()
-        title= str(self.ticket_obj.fields.summary).strip()
-        reporter= str(self.ticket_obj.fields.reporter.displayName).strip()
 
         if status.lower() in ["waiting for configuration", "configured"]:
-            icon, color_code = "🟢", "green"
             self.is_jira_valid= True
         else:
-            icon, color_code= "🔴", "red"
-            # self.is_jira_valid= False
+            self.is_jira_valid= False
 
-        self.jira_status.setText(
-            f"Jira Status: {icon}<br>Ticket: <b style='color:{color_code};'>{self.ticket_obj}</b>, "
-            f"Status: <b style='color:{color_code};'>{status}</b>"
-            f"<br>Title: {title}<br>Reporter: {reporter}")
+        validate_all_conditions(self) # Check if we can enable buttons
 
     except ValueError:
         error_message= ("ERROR! Unable to connect to JIRA server.")
@@ -63,13 +54,10 @@ def check_jira_status(self):
                         "valid JIRA ticket ID (e.g., CRF-12345)")
         handle_jira_error(self, "INVALID TICKET ID", error_message)
 
-    validate_all_conditions(self) # Check if we can enable buttons
-
 
 def add_jira_section(self):
     "Add the JIRA section to the GUI"
-    self.jira_status= QLabel(
-        "Jira Status: 🔴<br>Ticket: N/A, Status: N/A<br>Title: N/A<br>Reporter: N/A")
+    self.jira_status= QLabel()
     self.jira_layout= QHBoxLayout()
     self.jira_ticket= QLineEdit()
     self.jira_ticket.setPlaceholderText("i.e.: CRF-12345")
