+++ Misc Tools/Limits Version Control Tool/components/load_file.py
@@ -1,7 +1,7 @@
 import openpyxl
 from typing import List, Any
 from PyQt5.QtWidgets import QFileDialog, QLabel, QPushButton
-from components.misc import validate_all_conditions, create_separator
+from components.misc import validate_all_conditions
 
 
 def load_excel_raw(self) -> List[List[Any]]:
@@ -39,15 +39,13 @@ def open_file_dialog(self):
     "Open a file dialog to select a data file (xlsx)."
     self.fileName, _ = QFileDialog.getOpenFileName(self, "Select Data File", "", "Excel File (*.xlsx)")
     if ".xlsx" in self.fileName:
-        self.lbl_file_status.setText(f"File Status: 🟢<br>Loaded ({self.fileName.split('/')[-1]})")
         self.excel_file_data= load_excel_raw(self)
         self.is_file_loaded = True
-        validate_all_conditions(self)
+    validate_all_conditions(self)
 
 
 def clear_loaded_file(self):
     "Clear the loaded file"
-    self.lbl_file_status.setText("File Status: 🔴<br>No file selected")
     self.excel_file_data= None
     self.is_file_loaded= False
     validate_all_conditions(self) # Check if we can enable buttons
@@ -55,9 +53,8 @@ def clear_loaded_file(self):
 
 def add_file_select_section(self):
     "Add the 'File Selection' section to the GUI"
-    self.lbl_file_status = QLabel("File Status: 🔴<br>No file selected")
-    self.btn_load_file = QPushButton("Load File (.xlsx)")
+    self.lbl_file_status= QLabel()
+    self.btn_load_file= QPushButton("Load File (.xlsx)")
     self.btn_load_file.clicked.connect(lambda: open_file_dialog(self))
     self.layout.addWidget(self.lbl_file_status)
     self.layout.addWidget(self.btn_load_file)
-    self.layout.addWidget(create_separator(self))
