+++ Misc Tools/Limits Version Control Tool/components/misc.py
@@ -2,8 +2,6 @@
 
 import os.path
 from PyQt5.QtWidgets import QPushButton, QFrame
-from PyQt5.QtGui import QDesktopServices
-from PyQt5.QtCore import QUrl
 
 
 def get_user_directory(directory):
@@ -17,62 +15,86 @@ def get_user_directory(directory):
 def validate_all_conditions(self):
     """Checks if all criteria are met to enable the final action buttons."""
 
+    # Handle Google Log Status
     if self.is_logged_in:
-        self.btn_svn.setEnabled(True)
-        update_svn_button_style(self, enabled=True)
+        self.btn_login.setEnabled(False)
+        self.btn_logout.setEnabled(True)
+        self.lbl_auth_status.setText(f"CFA Google Account Log Status: 🟢"
+                                     f"<br>Logged In ({self.oauth_data['Email']})")
     else:
-        self.btn_svn.setEnabled(False)
-        update_svn_button_style(self, enabled= False)
+        self.btn_login.setEnabled(True)
+        self.btn_logout.setEnabled(False)
+        self.lbl_auth_status.setText("CFA Google Account Log Status: 🔴<br>Logged Out")
 
+    # Handle File Load Status
+    if self.is_file_loaded:
+        self.lbl_file_status.setText(f"File Status: 🟢<br>Loaded ({self.fileName.split('/')[-1]})")
+    else:
+        self.lbl_file_status.setText("File Status: 🔴<br>No file selected")
+
+    # Handle JIRA Status
+    try:
+        status= str(self.ticket_obj.fields.status).strip()
+        title= str(self.ticket_obj.fields.summary).strip()
+        reporter= str(self.ticket_obj.fields.reporter.displayName).strip()
+        if self.is_jira_valid:
+            self.jira_status.setText(
+                f"Jira Status: 🟢<br>Ticket: <b style='color:green;'>{self.ticket_obj}</b>, "
+                f"Status: <b style='color:green;'>{status}</b>"
+                f"<br>Title: {title}<br>Reporter: {reporter}")
+        else:
+            self.jira_status.setText(
+                f"Jira Status: 🔴<br>Ticket: <b style='color:red;'>{self.ticket_obj}</b>, "
+                f"Status: <b style='color:red;'>{status}</b>"
+                f"<br>Title: {title}<br>Reporter: {reporter}")
+    except AttributeError:
+        self.jira_status.setText(f"Jira Status: 🔴<br>Ticket: N/A, "
+                                 f"Status: N/A<br>Title: N/A<br>Reporter: N/A")
+
+    # Handle SVN Directory Status
+    if self.is_svn_valid:
+        self.svn_status.setText(f"SVN Status: 🟢<br>Directory: {self.svn_path}")
+    else:
+        self.svn_status.setText("SVN Status: 🔴<br>Directory: N/A")
+
+    # Enable Buttons
     if self.is_logged_in and self.is_jira_valid and self.is_file_loaded:
-        self.btn_merge.setEnabled(True)
-        update_merge_button_style(self, enabled=True)
+        enable_button(self.merge_btn)
     else:
-        self.btn_merge.setEnabled(False)
-        update_merge_button_style(self, enabled= False)
+        disable_button(self.merge_btn)
 
-
-def update_merge_button_style(self, enabled):
-    if enabled:
-        self.btn_merge.setStyleSheet("background-color: #28a745; color: white; font-size: 18px; "
-                                    "font-weight: bold; padding: 15px; border-radius: 5px;")
+    if self.is_svn_valid:
+        enable_button(self.svn_pull_btn)
     else:
-        self.btn_merge.setStyleSheet("background-color: #d3d3d3; color: #888888; font-size: 12px; "
-                                    "font-weight: bold; padding: 12px; border-radius: 5px;")
+        disable_button(self.svn_pull_btn)
 
-
-def update_svn_button_style(self, enabled):
-    if enabled:
-        self.btn_svn.setStyleSheet("background-color: #28a745; color: white; font-size: 18px; "
-                                   "font-weight: bold; padding: 15px; border-radius: 5px;")
+    if self.is_logged_in and self.is_jira_valid and self.is_svn_valid:
+        enable_button(self.svn_save_btn)
     else:
-        self.btn_svn.setStyleSheet("background-color: #d3d3d3; color: #888888; font-size: 12px; "
-                                   "font-weight: bold; padding: 12px; border-radius: 5px;")
+        disable_button(self.svn_save_btn)
+    
+    if self.is_logged_in:
+        enable_button(self.save_btn)
+    else:
+        disable_button(self.save_btn)
 
 
-def update_master_button_layout(self):
-    self.btn_master.setStyleSheet("background-color: #28a745; color: white; font-size: 12px; "
-                                "font-weight: bold; padding: 10px; border-radius: 5px;")
+def enable_button(button):
+    button.setEnabled(True)
+    button.setStyleSheet("background-color: #28a745; color: white; font-size: 12px; "
+                         "font-weight: bold; padding: 12px; border-radius: 5px;")
 
 
-def open_sheets_master():
-    url = QUrl("https://docs.google.com/spreadsheets/d/15rRk5JAMWXBGiKTly4aP0cUuFE1qECZe01tNESSKXBo/edit?usp=sharing")
-    QDesktopServices.openUrl(url)
+def disable_button(button):
+    button.setEnabled(False)
+    button.setStyleSheet("background-color: #d3d3d3; color: #888888; font-size: 12px; "
+                         "font-weight: bold; padding: 12px; border-radius: 5px;")
 
 
 def create_separator(self):
     line= QFrame(); line.setFrameShape(QFrame.HLine)
     line.setFrameShadow(QFrame.Sunken)
     return line
-
-
-def add_sheet_master_btn(self):
-    "Add the 'Open Sheets Master' button to the GUI"
-    self.btn_master= QPushButton("Open Sheets Master")
-    self.btn_master.setEnabled(True)
-    self.btn_master.clicked.connect(lambda: open_sheets_master())
-    self.layout.addWidget(self.btn_master)
-    update_master_button_layout(self)
 
 
 def add_exit_btn(self):
